preserve return values

+++ cubi_tk/isa_tpl/__init__.py
@@ -258,21 +258,27 @@ def run(args, parser, subparser):  # pragma: nocover
     if not args.tpl:  # pragma: nocover
         return run_nocmd(args, parser, subparser)
     else:
-        args.isa_tpl_cmd(args, parser, subparser)
+        status = args.isa_tpl_cmd(args, parser, subparser)
 
         # output validation
-        logger.info("Running AltamISA validator:")
-        i_files = Path(args.output_dir).rglob("i_*")
-        args.show_duplicate_warnings = False
-        warnings.filterwarnings("error", category=altamisa.exceptions.CriticalIsaValidationWarning)
-        for i in i_files:
-            try:
-                with i.open() as i_file:
-                    args.input_investigation_file = i_file
-                    altamisa.apps.isatab_validate.run(args)
-            except (
-                altamisa.exceptions.ParseIsatabException,
-                altamisa.exceptions.CriticalIsaValidationWarning,
-            ):
-                shutil.rmtree(args.output_dir)
-                raise
+        if not status:
+            logger.info("Running AltamISA validator:")
+            i_files = Path(args.output_dir).rglob("i_*")
+            args.show_duplicate_warnings = False
+            warnings.filterwarnings(
+                "error", category=altamisa.exceptions.CriticalIsaValidationWarning
+            )
+            for i in i_files:
+                try:
+                    with i.open() as i_file:
+                        args.input_investigation_file = i_file
+                        altamisa.apps.isatab_validate.run(args)
+                except (
+                    altamisa.exceptions.ParseIsatabException,
+                    altamisa.exceptions.CriticalIsaValidationWarning,
+                ):
+                    shutil.rmtree(args.output_dir)
+                    raise
+            return 0
+        else:
+            return status
